put labelstudio where it was
